Version 1.12-1.8.5 beta uses this code
Players now break blocks (3*noOfBreaks) blocks ahead of them, to
compromise for packet latency. A getter for the distance ahead to break
was added to IBreakCapabality

Integrated CommonBreakingMethods into BreakSystem

A client tick listener sends a message with new player speed to the
server if the player's speed varies by more than 0.01. The message is
used to update the player's IPlayerBreakCapability. The player's speed
is obtained on the server from the capability instead of the motionX/Y/Z
values

Removed EntityFollower

Using ModelLoader and ModelRegistryEvents to register models, instead of
preInit and ItemModelMesher

onCrash is now a boolean. Returning true will stop the bounding box
being offset any further in the break check for that Entity for that
tick

Multiplayer compatible

Updated credits

Improved sugar cauldron model (added legs)