Add Fragilitator item and TileEntityBlockMadeFragile
When right-clicking a block, if the block has no TileEntity, the Fragilitator should place a TileEntityBlockMadeFragile, which should give the block the fragile behaviour specified in fragileglassft_blocklist.cfg. The Tile Entity destroys itself when a check reveals simply an air block in its position.
Separate fragile data HashMap for TileEntity-less blocks